fix(tests): split test file to comply with LOC ceiling
Split test_manager_scene_capabilities.py into three focused modules:

- test_manager_scene_capabilities_observation.py (303 lines)
  Tests for SceneStatus properties and observation methods

- test_manager_scene_capabilities_cleanup.py (146 lines)
  Tests for cleanup methods delegation and error handling

- test_manager_scene_capabilities_integration.py (71 lines)
  Tests for edge cases and integration scenarios

All files now comply with the 400-line maximum LOC ceiling for test files.
All 28 tests pass with no type errors or lint issues.

Fixes CI failure in PR #616.
